% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ath_featureLength.R
\docType{data}
\name{ath_featureLength}
\alias{ath_featureLength}
\title{Example - vector with gene length}
\format{Name vector: length (values) of genes (names) of Arabidopsis thaliana}
\source{
\url{plants.ensembl.org}
}
\usage{
ath_featureLength
}
\description{
Length of Arabidopsis thaliana genes (TAIR10) obtained with the following code:
}
\details{
\code{library(biomaRt)}

\code{ath <- useMart('plants_mart', host = "plants.ensembl.org", dataset = "athaliana_eg_gene")}

\code{gene_start_end = getBM(attributes = c('ensembl_gene_id', 'start_position', 'end_position'), mart = ath)}

\code{featureLength <- gene_start_end$end_position - gene_start_end$start_position}

\code{names(featureLength) <- gene_start_end$ensembl_gene_id}
}
\keyword{datasets}
