% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Counts_to_tpm.R
\name{Counts_to_tpm}
\alias{Counts_to_tpm}
\title{Counts_to_tpm
This function was modified from a gist from Slowkow, we do not calculate the effective length
With the matrix of counts and the size of the genes/transcripts, it calculates the TPM}
\usage{
Counts_to_tpm(counts, featureLength)
}
\arguments{
\item{counts}{matrix of genes/features (in the rows) by samples (in the columns)}

\item{featureLength}{a named numeric vector with the length of each gene/feature}
}
\value{
list with: a data frame with TPM of genes (rows) per sample (columns)
and a vector with the names of the genes/features for which the TPM was calculated
}
\description{
Counts_to_tpm
This function was modified from a gist from Slowkow, we do not calculate the effective length
With the matrix of counts and the size of the genes/transcripts, it calculates the TPM
}
\references{
https://gist.github.com/slowkow/c6ab0348747f86e2748b
}
