% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DAFS.R
\name{DAFS}
\alias{DAFS}
\title{DAFS}
\usage{
DAFS(tpm)
}
\arguments{
\item{tpm}{Data frame of TPM of genes (rows) per sample (columns)}
}
\value{
a vector with threshold for noise/true expression for each sample (columns from counts) in log2
}
\description{
Modified from George and Chang (2014)
The function normally requires two inputs: data and name.
We do not export the result, so we do not need a 'name' argument.
}
\references{
doi:10.1186/1471-2105-15-92
}
\author{
George and Chang (2014)
}
