% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_selection.R
\name{gene_selection}
\alias{gene_selection}
\title{Uses average TPM values and the covariance of TPM values to select reference genes from RNAseq data}
\usage{
gene_selection(countsToTpm_result, dafs_result, top_genes = 0.5)
}
\arguments{
\item{countsToTpm_result}{Result of the counts_to_tpm function or data frame with other expression unit (FPKM, RPKM, CPM)}

\item{dafs_result}{Result of DAFS fucntion: vector with threshold for noise/true expression for each sample (columns from tpm data frame) in log2}

\item{top_genes}{Percentage of genes (left after filtering) to be selected as references, default is 0.5\%.}
}
\value{
Data frame with genes as rows and two columns: Mean (average TPM) and Covariance.
}
\description{
If counts_to_tpm and DAFS functions were already computed, this function will use their results to select the genes with lowest covariance, among those considered as expressed according to DAFS, as references.
}
\details{
If there is no interest on keeping the results from DAFS or the TPM values for all genes in each sample, run customReferences instead.
}
