% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Counts_to_tpm.R
\name{Counts_to_tpm}
\alias{Counts_to_tpm}
\title{Transforms count data into Transcripts Per Million (TPM) data}
\usage{
Counts_to_tpm(counts, featureLength)
}
\arguments{
\item{counts}{Data frame genes/features (in the rows) by samples (in the columns).
Rownames must be in the names of the features and in the same format as the names of the featureLength vector}

\item{featureLength}{Named numeric vector with the length of each gene/feature.
Names must be the names of the features and in the same format as the rownames of the data frame counts.}
}
\value{
List with: a data frame with TPM of genes (rows) per sample (columns)
and a vector with the names of the genes/features for which the TPM was calculated
}
\description{
With the matrix of counts and the size of the genes/transcripts, it calculates the TPM.
}
\details{
This function was modified from a gist from Slowkow (https://gist.github.com/slowkow/c6ab0348747f86e2748b).

Here, we do not calculate the effective length.
}
\examples{
data("sample_counts"); data("ath_featureLength")
Counts_to_tpm(counts = sample_counts, featureLength = ath_featureLength)

}
\references{
https://gist.github.com/slowkow/c6ab0348747f86e2748b
}
